+++ src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css';
-import SignIn from './pages/SignIn';
-import SignUp from './pages/SignUp';
+import SignIn from './pages/SignInPage';
+import SignUp from './pages/SignUpPage';
 
 function App() {
   return (
